backend/requirements.txt
# Core Flask and Web Server
flask>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Machine Learning and Similarity Search
numpy>=1.24.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Embedding Models (BGE-M3)
FlagEmbedding>=1.2.10
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.36.0

# Vector Database
pgvector>=0.2.5

# Natural Language Processing
nltk>=3.8.1

# HTTP Requests
requests>=2.31.0

# Retry Logic
tenacity>=8.2.3

# Rate Limiting
Flask-Limiter>=3.5.0
